cloud-functions/ingest-drive-video/requirements.txt
functions-framework==3.*
google-cloud-storage
google-api-python-client
google-auth
requests